+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Batteries [1.0.5]
+- Natural spawning of batteries can now be disabled by setting the batteryRarity config value to 0
+- Batteries can now be configured to appear in the shop with a price as defined by the batteryShopValue config value
+- The charging coil can now be disabled using the disableChargingCoil config value to make batteries the only charge source
+
 ## Batteries [1.0.4]
 - Fixed batteries not saving, despawning, selling for any value, and other shenanigans due to not being marked as scrap
 	- I originally wanted for batteries to be like keys, but making it function as a full scrap seems to fix several issues